requirements.txt
click==8.0.3
matplotlib==3.5.0
numpy==1.21.4
Pillow==9.0.0
python-dotenv==0.19.2
scikit_learn==1.0.2
setuptools==59.4.0
torch==1.10.1
torchvision==0.11.2
